Fix remove_tag when links are disabled
When links are disabled, the li.querySelector('span') returns the outer
span, which includes the tag text as well as the close 'x'.  This causes
this._tags to never change because we end up comparing 'foox' against
'foo'.

This commit fixes this by putting the tag text into a span with the
'tagspan' class, so when we call li.querySelector, we select based on
this class name to avoid including the close 'x'.